fix: types for config

+++ src/models/config.py
@@ -75,8 +75,10 @@ class RovConfig(CamelCaseModel):
     )
     fluid_type: FluidType = FluidType.SALTWATER
     thruster_pin_setup: ThrusterPinSetup = ThrusterPinSetup(
-        identifiers=np.array([0, 1, 2, 3, 4, 5, 6, 7], dtype=int),
-        spin_directions=np.array([1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], dtype=float),
+        identifiers=np.array([0, 1, 2, 3, 4, 5, 6, 7], dtype=np.int8),
+        spin_directions=np.array(
+            [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0], dtype=np.float32
+        ),
     )
     thruster_allocation: NDArray[Shape["8, 8"], np.float32] = np.array(  # pyright: ignore[reportInvalidTypeForm]
         (
